+++ RELEASES.md
@@ -33,6 +33,16 @@ Upcoming
 - Workaround to avoid NPE after a full reload
 - Avoid warning when ShowFlowDiagram REST listener is unavailable
 - Support document(), xsl:import, and xsl:include with config in database
+- Improve error handling TransformerPools and TransformerFactories
+- Add support for configuration hierarchies. Configurations can define a parent configuration via the property 'configurations.<config-name>.parentConfig=<parentConfig-name>'
+- Fix property file classloading with suffix files.The classloader will search and consolidate all properties found in the classloader chain
+- Introduction of the RestSender
+
+
+
+### Non backwards compatible changes
+
+- ADDITIONAL.PROPERTIES.FILE.SUFFIX has been made obsolete. Users are adviced to remove suffix files as soon as possible.